// src/app/model/diplome.ts
export class Diplome {
    id:number=0;
    name: string='';
    niveau: string='';
    domaineEtudes: string='';
    specialisation: string='';
    institution: string='';
    pays: string='';
    dateObtention: Date | null = null;
    duree: number=0;
    notes: string='';
    detenteur: string='';
}